Update for front 3

+++ src/main/java/com/example/mycli/repository/UserEntityRepo.java
@@ -15,6 +15,6 @@ public interface UserEntityRepo extends JpaRepository<UserEntity, Long> {
 
     Optional<UserEntity> findByVerificationCode(String verificationCode);
 
-    List<UserEntity> findAllByAuthdata_RoleEntity_Id(Long id);
+    List<UserEntity> findAllByAuthdata_RoleEntity_Id(int id);
 
 }
